pyproject.toml
[tool.poetry]
authors = ["eschalk <etienne.schalk@gmail.com>"]
description = ""
name = "advent-of-code"
packages = [{ include = "advent_of_code" }]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
click = "^8.1.7"
python = "^3.12"
scikit-image = "^0.22.0"
sympy = "^1.12"
xarray = "^2023.12.0"
zarr = "^2.16.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
graphviz = "^0.20.1"
hvplot = "^0.9.1"
ipykernel = "^6.27.1"
matplotlib = "^3.8.2"
notebook = "^7.0.6"
pre-commit = "^3.6.0"
pyright = "^1.1.345"
pytest = "^7.4.4"
pytest-integration-mark = "^0.2.0"
pytest-optional-tests = "^0.1.1"
pytest-xdist = "^3.5.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
# mkdocs-material = "^9.5.9"
# mkdocstrings = {extras = ["python"], version = "^0.24.0"}
furo = "^2024.1.29"
myst-parser = "^2.0.0"
nbsphinx = "^0.9.3"
sphinx-autoapi = "^3.0.0"
sphinx-autobuild = "^2024.2.4"
sphinx-mdinclude = "^0.5.3"
sphinx-serve = "^1.0.1"
sphinxcontrib-youtube = "^1.4.1"
sphinxemoji = "^0.3.1"


[tool.poetry.group.visu2d]
optional = true

[tool.poetry.group.visu2d.dependencies]
pygame-ce = "^2.4.0"
pymunk = "^6.6.0"

# Actually unused yet
[tool.poetry.group.visu3d]
optional = true

[tool.poetry.group.visu3d.dependencies]
panda3d = "^1.10.14"
ursina = "^6.1.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.pyright]
# ...
exclude = ["tests/"] # Be lenient for tests.
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownVariableType = false
venv = "advent-of-code-CEbbwEIT-py3.12"
venvPath = "/home/tselano/.cache/pypoetry/virtualenvs/"

[tool.pytest.ini_options]
addopts = "-m 'not slow'"
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
